refactor: crea páginas para pruebas
