+++ src/pages/DarkTheme.module.scss
@@ -1,0 +1,26 @@
+:root {
+  --background-color: #e3e3e3;
+  --text-color: #222;
+}
+
+.dark-theme {
+  --background-color: #222;
+  --text-color: #ccc;
+}
+
+.container {
+  padding: 60px;
+
+  transition: background-color 0.5s, color 0.5s;
+
+  background-color: var(--background-color);
+  color: var(--text-color);
+}
+
+.title {
+  color: red;
+}
+
+.gsap {
+  display: inline-block;
+}
