+++ src/pages/index.module.scss
@@ -1,25 +1,3 @@
-:root {
-  --background-color: #e3e3e3;
-  --text-color: #222;
-}
-
-.dark-theme {
-  --background-color: #222;
-  --text-color: #ccc;
-}
-
-.wrapper {
-  background-color: var(--background-color);
-  color: var(--text-color);
-}
-
-.container {
-  transition: background-color 0.5s, color 0.5s;
-
-  background-color: var(--background-color);
-  color: var(--text-color);
-}
-
 .title {
   color: red;
 }
